fix some bugs

+++ src/train/psgd/PriorityHolder.h
@@ -7,9 +7,9 @@ class PriorityHolder{
 public:
 	std::vector<std::pair<float, unsigned>> priority;
 	virtual void init(const size_t size);
-	virtual float get(const int id, const unsigned ver);
-	virtual void set(const int id, const unsigned ver, const float prio);
-	virtual void update(const int id, const unsigned ver, const float prio);
+	virtual float get(const size_t id, const unsigned ver);
+	virtual void set(const size_t id, const unsigned ver, const float prio);
+	virtual void update(const size_t id, const unsigned ver, const float prio);
 };
 
 // p_n = p_o * exp(a * n)
@@ -17,9 +17,9 @@ class PriorityHolderExpLinear: public PriorityHolder {
 	std::vector<float> factor;
 public:
 	virtual void init(const size_t size);
-	virtual float get(const int id, const unsigned ver);
-	virtual void set(const int id, const unsigned ver, const float prio);
-	virtual void update(const int id, const unsigned ver, const float prio);
+	virtual float get(const size_t id, const unsigned ver);
+	virtual void set(const size_t id, const unsigned ver, const float prio);
+	virtual void update(const size_t id, const unsigned ver, const float prio);
 };
 
 // p_n = p_o * exp( (a*n + b) * n)
@@ -28,8 +28,8 @@ class PriorityHolderExpTwice : public PriorityHolder {
 	std::vector<std::pair<float, float>> factor;
 public:
 	virtual void init(const size_t size);
-	virtual float get(const int id, const unsigned ver);
-	virtual void set(const int id, const unsigned ver, const float prio);
-	virtual void update(const int id, const unsigned ver, const float prio);
+	virtual float get(const size_t id, const unsigned ver);
+	virtual void set(const size_t id, const unsigned ver, const float prio);
+	virtual void update(const size_t id, const unsigned ver, const float prio);
 };
